feat(app): add ProfileDropdown
